Add user following and follower relations
This commit adds support for managing, tracking, and querying for user follower and following relationships. The relations are saved in a GORM-enabled database and include views to list all followed users, users who follow a given user, and mutual followers ("friends"). The commit also includes specific error handling for records not found, which will be useful for user interface messages.

The relation structure, all database initializations, and API router updates are included. Updates also include Thrift IDLs and compiled Kitex files.

The changes in this commit are essential for building a functional social media app, as following relationships are at the core of any social media platform.